+++ test/Sdk.Generator.Tests/FunctionMetadataProviderGeneratorTests/HttpTriggerTests.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using System.Collections.Generic;
 using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker.Sdk.Generators;
@@ -65,11 +66,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -150,11 +152,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -239,11 +242,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace MyCompany.MyProject.MyApp
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -285,12 +289,18 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                     }
                 }
                 """;
+                // override the namespace value for generated types using msbuild property.
+                var buildPropertiesDict = new Dictionary<string, string>()
+                {
+                    {  Constants.BuildProperties.GeneratedCodeNamespace, "MyCompany.MyProject.MyApp"}
+                };
 
                 await TestHelpers.RunTestAsync<FunctionMetadataProviderGenerator>(
                     _referencedExtensionAssemblies,
                     inputCode,
                     expectedGeneratedFileName,
-                    expectedOutput);
+                    expectedOutput,
+                    buildPropertiesDictionary: buildPropertiesDict);
             }
         }
     }