+++ test/Sdk.Generator.Tests/FunctionMetadataProviderGeneratorTests/NestedTypesTest.cs
@@ -71,11 +71,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {
@@ -157,11 +158,12 @@ namespace Microsoft.Azure.Functions.SdkGeneratorTests
                 using System.Collections.Immutable;
                 using System.Text.Json;
                 using System.Threading.Tasks;
+                using Microsoft.Azure.Functions.Worker;
                 using Microsoft.Azure.Functions.Worker.Core.FunctionMetadata;
                 using Microsoft.Extensions.DependencyInjection;
                 using Microsoft.Extensions.Hosting;
 
-                namespace Microsoft.Azure.Functions.Worker
+                namespace TestProject
                 {
                     public class GeneratedFunctionMetadataProvider : IFunctionMetadataProvider
                     {