+++ test/Worker.Extensions.Rpc.Tests/RpcServiceCollectionExtensionsTests.cs
@@ -4,12 +4,11 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using Moq;
 using Xunit;
 
 namespace Microsoft.Azure.Functions.Worker.Extensions.Rpc.Tests
 {
-    public class WorkerRpcStartupTests
+    public class RpcServiceCollectionExtensionsTests
     {
         [Fact]
         public void Configure_AddsFunctionsGrpcOptions()
@@ -24,11 +23,7 @@ namespace Microsoft.Azure.Functions.Worker.Extensions.Rpc.Tests
 
             ServiceCollection services = new();
             services.AddSingleton((IConfiguration)configBuilder.Build());
-            IFunctionsWorkerApplicationBuilder builder = Mock.Of<IFunctionsWorkerApplicationBuilder>(
-                m => m.Services == services);
-
-            WorkerRpcStartup startup = new();
-            startup.Configure(builder);
+            services.AddWorkerRpc();
 
             IServiceProvider sp = services.BuildServiceProvider();
             IOptions<FunctionsGrpcOptions> options = sp.GetService<IOptions<FunctionsGrpcOptions>>();