+++ sdk/release_notes.md
@@ -14,4 +14,8 @@
 
 ### Microsoft.Azure.Functions.Worker.Sdk.Generators <version>
 
-- <entry>
+- Parse named arguments by type (#1877)
+- Refactor source gen to walk dependent assemblies (#1896)
+- Add diagnostic descriptor logs for parsing binding arguments in source gen (#1882)
+- Use project root namespace for generated types (#1158)
+